Diploma Engineering Course Seeder Add & Controller Code Update

+++ app/Http/Controllers/Admin/DiplomaEngineeringCourseController.php
@@ -34,7 +34,7 @@ class DiplomaEngineeringCourseController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'image' =>['required', 'image', 'max:2000'],
+            'image' => ['required', 'image', 'max:2000'],
             'name' => ['required', 'max:255', 'unique:diploma_engineering_courses,name'],
             'month' => ['required', 'max:255'],
             'duration' => ['required', 'max:255'],
@@ -49,11 +49,11 @@ class DiplomaEngineeringCourseController extends Controller
         if ($request->file('image')) {
             $image = $request->file('image');
             $manager = new ImageManager(new Driver());
-            $name_gen = hexdec(uniqid()).'.'.$image->getClientOriginalExtension();
+            $name_gen = hexdec(uniqid()) . '.' . $image->getClientOriginalExtension();
             $img = $manager->read($image);
-            $img = $img->resize(1110,555);
-            $img->toPng()->save(base_path('public/uploads/diploma_engineering_course_image/'.$name_gen));
-            $save_url = 'uploads/diploma_engineering_course_image/'.$name_gen;
+            $img = $img->resize(1110, 555);
+            $img->toPng()->save(base_path('public/uploads/diploma_engineering_course_image/' . $name_gen));
+            $save_url = 'uploads/diploma_engineering_course_image/' . $name_gen;
 
             $diplomaEngineeringCourse = new DiplomaEngineeringCourse();
             $diplomaEngineeringCourse->image = $save_url;
@@ -96,8 +96,8 @@ class DiplomaEngineeringCourseController extends Controller
     public function update(Request $request, string $id)
     {
         $request->validate([
-            'image' =>['nullable', 'image', 'max:2000'],
-            'name' => ['required', 'max:255', 'unique:diploma_engineering_courses,name,'.$id],
+            'image' => ['nullable', 'image', 'max:2000'],
+            'name' => ['required', 'max:255', 'unique:diploma_engineering_courses,name,' . $id],
             'month' => ['required', 'max:255'],
             'duration' => ['required', 'max:255'],
             'course_fee' => ['required', 'max:255'],
@@ -111,11 +111,11 @@ class DiplomaEngineeringCourseController extends Controller
         if ($request->file('image')) {
             $image = $request->file('image');
             $manager = new ImageManager(new Driver());
-            $name_gen = hexdec(uniqid()).'.'.$image->getClientOriginalExtension();
+            $name_gen = hexdec(uniqid()) . '.' . $image->getClientOriginalExtension();
             $img = $manager->read($image);
-            $img = $img->resize(1110,555);
-            $img->toPng()->save(base_path('public/uploads/diploma_engineering_course_image/'.$name_gen));
-            $save_url = 'uploads/diploma_engineering_course_image/'.$name_gen;
+            $img = $img->resize(1110, 555);
+            $img->toPng()->save(base_path('public/uploads/diploma_engineering_course_image/' . $name_gen));
+            $save_url = 'uploads/diploma_engineering_course_image/' . $name_gen;
 
             $diplomaEngineeringCourse = DiplomaEngineeringCourse::findOrFail($id);
             $diplomaEngineeringCourse->image = $save_url;
@@ -130,13 +130,13 @@ class DiplomaEngineeringCourseController extends Controller
             $diplomaEngineeringCourse->status = $request->status;
             $diplomaEngineeringCourse->save();
 
-            if (file_exists($oldImage)) {
+            if (file_exists($oldImage !== 'frontend/images/courses/course-1.jpg' || $oldImage !== 'frontend/images/courses/course-2.jpg' || $oldImage !== 'frontend/images/courses/course-3.jpg' || $oldImage !== 'frontend/images/courses/course-4.jpg' || $oldImage !== 'frontend/images/courses/course-5.jpg' || $oldImage !== 'frontend/images/courses/course-6.jpg')) {
                 unlink($oldImage);
             }
 
             toastr()->success('Updated Successfully');
             return redirect()->route('admin.diploma-engineering-course.index');
-        }else{
+        } else {
             $diplomaEngineeringCourse = DiplomaEngineeringCourse::findOrFail($id);
             $diplomaEngineeringCourse->name = $request->name;
             $diplomaEngineeringCourse->slug = Str::slug($request->name);
@@ -160,9 +160,12 @@ class DiplomaEngineeringCourseController extends Controller
     public function destroy(string $id)
     {
         $diplomaEngineeringCourse = DiplomaEngineeringCourse::findOrFail($id);
-        unlink($diplomaEngineeringCourse->image);
-        $diplomaEngineeringCourse->delete();
 
+        if (file_exists($diplomaEngineeringCourse !== 'frontend/images/courses/course-1.jpg' || $diplomaEngineeringCourse !== 'frontend/images/courses/course-2.jpg' || $diplomaEngineeringCourse !== 'frontend/images/courses/course-3.jpg' || $diplomaEngineeringCourse !== 'frontend/images/courses/course-4.jpg' || $diplomaEngineeringCourse !== 'frontend/images/courses/course-5.jpg' || $diplomaEngineeringCourse !== 'frontend/images/courses/course-6.jpg')) {
+            unlink($diplomaEngineeringCourse->image);
+        }
+
+        $diplomaEngineeringCourse->delete();
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
 }
